requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
SQLAlchemy>=2.0
psycopg[binary]>=3.1
python-dotenv>=1.0
pydantic>=2.0
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
